10.20

+++ client/src/client.py
@@ -1,18 +1,18 @@
 import socket
 import sys
 from enum import Enum
+import time
+from pathlib import Path
+import os
 
 DEFAULT_IP = '127.0.0.1'
 DEFAULT_PORT = 21
 BUFFER_SIZE = 8192
 
-# TODO 如果没PASV或PORT要报错
-# TODO 连服务器，测试大文件传输
 # TODO 纠错
 # TODO 选做，图形化界面
 # TODO 选做，断线重连
 # TODO 传输文件开启新线程
-# TODO 少几个命令
 
 class Method(Enum):
     NOTHING = 1
@@ -24,6 +24,7 @@ class Data_connection_method:
         self.method: Method = Method.NOTHING
         self.ip_address: str = ""
         self.port_number: int = 0
+        self.data_sock = 0
 
 def parse_arguments():
     ip_address = DEFAULT_IP
@@ -52,7 +53,8 @@ def send_command(sock: socket.socket, command):
     """
     向server发送一次请求并获得一次预期response.
     """
-    sock.sendall(command.encode())
+    message = command + "\r\n"
+    sock.sendall(message.encode())
     return receive_response(sock)
 
 def receive_response(sock: socket.socket):
@@ -74,37 +76,37 @@ def receive_response(sock: socket.socket):
             if len(last_line) >= 4 and last_line[3] == ' ':
                 break
 
-    print("RESPONSE:\n", response.strip())
+    print(response.strip(), flush=True)
     return last_line.strip()
 
+def get_base_path():
+    if getattr(sys, '_MEIPASS', False):
+        return Path(sys.executable).parent
+    else:
+        return Path(__file__).parent
+    
 def retrieve_file(sock: socket.socket, filename: str, data_connection_method: Data_connection_method):
-    local_save_path = input("Input the local path to save the downloaded file: ")
+    # current_directory = "/Users/qiaoshenyu/Desktop/大三上/计网/lhw1/ForStudents/autograde/autograde_client"
+    current_directory = get_base_path().resolve()
+    # current_directory = "/home/shenyu/projects/ftp/client"
+    
+    local_save_path = os.path.join(current_directory, filename)
     response = send_command(sock, f"RETR {filename}")
 
     if response.startswith("550 "):
         return False
-    elif response.startswith("150 "):
+    elif response.startswith("150 ") or response.startswith("125 "):
         if data_connection_method.method == Method.PORT:
-            data_ip_address = data_connection_method.ip_address
-            data_port_number = data_connection_method.port_number
-
-            data_sock_listen = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            data_sock_listen.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-
-            data_sock_listen.bind((data_ip_address, data_port_number))
-            data_sock_listen.listen(1)
-
-            data_sock, addr = data_sock_listen.accept()
-            print("Data connection established.")
+            data_sock, addr = data_connection_method.data_sock.accept()
         elif data_connection_method.method == Method.PASV:
             data_ip_address = data_connection_method.ip_address
             data_port_number = data_connection_method.port_number
 
             data_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
             data_sock.connect((data_ip_address, data_port_number))
-            print("Data connection established.")
         else:
             print("Data connection must be established befoere transporting data.")
+            receive_response(sock)
             return False
         
         with open(local_save_path, 'wb') as f:
@@ -113,15 +115,14 @@ def retrieve_file(sock: socket.socket, filename: str, data_connection_method: Da
                 if not data:
                     break
                 f.write(data)
-                print(f"Received {len(data)} bytes")
         
         data_sock.close()
         if data_connection_method == Method.PORT:
-            data_sock_listen.close()
+            data_connection_method.data_sock.close()
 
         response = receive_response(sock)
         if response.startswith("226 "):
-            print(f"File {filename} retrieved successfully.")
+            pass
         else:
             print("Something went wrong while storing.")
 
@@ -129,51 +130,45 @@ def retrieve_file(sock: socket.socket, filename: str, data_connection_method: Da
     else:
         pass
 
-def store_file(sock: socket.socket, filename: str, data_connection_method: Data_connection_method):
-    local_file_name = input("Input the local file to upload: ")
-    response = send_command(sock, f"STOR {filename}")
+def store_file(sock: socket.socket, filename: str, data_connection_method: Data_connection_method) -> bool:
+    if not os.path.isfile(filename):
+        print(f"File '{filename}' does not exist.")
+        return False
 
-    if response.startswith("550 "):
+    base_filename = os.path.basename(filename)
+
+    response = send_command(sock, f"STOR {base_filename}")
+
+    if response.startswith("451 "):
         return False
     elif response.startswith("150 "):
         if data_connection_method.method == Method.PORT:
-            data_ip_address = data_connection_method.ip_address
-            data_port_number = data_connection_method.port_number
-
-            data_sock_listen = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            data_sock_listen.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-
-            data_sock_listen.bind((data_ip_address, data_port_number))
-            data_sock_listen.listen(1)
-
-            data_sock, addr = data_sock_listen.accept()
-            print("Data connection established.")
+            data_sock, addr = data_connection_method.data_sock.accept()
         elif data_connection_method.method == Method.PASV:
             data_ip_address = data_connection_method.ip_address
             data_port_number = data_connection_method.port_number
 
             data_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
             data_sock.connect((data_ip_address, data_port_number))
-            print("Data connection established.")
         else:
             print("Data connection must be established befoere transporting data.")
+            receive_response(sock)
             return False
         
-        with open(local_file_name, 'rb') as f:
+        with open(filename, 'rb') as f:
             while True:
                 data = f.read(BUFFER_SIZE)
                 if not data:
                     break
                 data_sock.sendall(data)
-                print(f"Sent {len(data)} bytes")
         
         data_sock.close()
         if data_connection_method.method == Method.PORT:
-            data_sock_listen.close()
+            data_connection_method.data_sock.close()
         
         response = receive_response(sock)
         if response.startswith("226 "):
-            print(f"File {filename} stored successfully.")
+            pass
         else:
             print("Something went wrong while storing.")
         
@@ -191,16 +186,7 @@ def list_files(sock: socket.socket, data_connection_method: Data_connection_meth
     response = send_command(sock, "LIST")
     if response.startswith("150 "):
         if data_connection_method.method == Method.PORT:
-            data_ip_address = data_connection_method.ip_address
-            data_port_number = data_connection_method.port_number
-
-            data_sock_listen = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            data_sock_listen.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-
-            data_sock_listen.bind((data_ip_address, data_port_number))
-            data_sock_listen.listen(1)
-
-            data_sock, addr = data_sock_listen.accept()
+            data_sock, addr = data_connection_method.data_sock.accept()
             print("Data connection established.")
         elif data_connection_method.method == Method.PASV:
             data_ip_address = data_connection_method.ip_address
@@ -211,10 +197,9 @@ def list_files(sock: socket.socket, data_connection_method: Data_connection_meth
             print("Data connection established.")
         else:
             print("Data connection must be established before retrieving the file list.")
-            response = receive_response(sock)
+            receive_response(sock)
             return False
         
-        print("Receiving file list:")
         file_list = ''
         while True:
             data = data_sock.recv(BUFFER_SIZE).decode()
@@ -225,7 +210,7 @@ def list_files(sock: socket.socket, data_connection_method: Data_connection_meth
         print(file_list)
         data_sock.close()
         if data_connection_method.method == Method.PORT:
-            data_sock_listen.close()
+            data_connection_method.data_sock.close()
 
         response = receive_response(sock)
         if response.startswith("226 "):
@@ -242,28 +227,28 @@ def change_directory(sock: socket.socket, directory: str):
     response = send_command(sock, f"CWD {directory}")
     
     if response.startswith("250 "):
-        print(f"Changed directory to {directory}")
+        pass
     else:
         print(f"Failed to change directory to {directory}: {response}")
 
 def print_working_directory(sock: socket.socket):
     response = send_command(sock, "PWD")
     if response.startswith("257 "):
-        print(f"Current directory: {response}")
+        pass
     else:
         print(f"Failed to get working directory: {response}")
 
 def make_directory(sock: socket.socket, directory: str):
     response = send_command(sock, f"MKD {directory}")
     if response.startswith("257 "):
-        print(f"Directory {directory} created successfully.")
+        pass
     else:
         print(f"Failed to create directory {directory}: {response}")
 
 def remove_directory(sock: socket.socket, directory: str):
     response = send_command(sock, f"RMD {directory}")
     if response.startswith("250 "):
-        print(f"Directory {directory} removed successfully.")
+        pass
     else:
         print(f"Failed to remove directory {directory}: {response}")
 
@@ -277,51 +262,70 @@ def main():
         
         # 读取欢迎消息
         response = sock.recv(BUFFER_SIZE).decode()
-        print("RESPONSE:\n", response)
+        time.sleep(0.1)
+        print(response, flush=True)
 
         logged_in = False
         # 登录逻辑
         while not logged_in:
-            user_command = input("Enter USER command (e.g., USER anonymous): ")
+            user_command = input()
             if user_command.startswith("USER"):
                 response = send_command(sock, user_command)
                 
-                # 使用返回码判断是否需要输入密码
                 if response.startswith("331 "):  # 331 表示需要密码
-                    email = input("Enter your email as password (e.g., PASS your_email@example.com): ")
-                    response = send_command(sock, f'PASS {email}')
-                    
-                    # 使用返回码判断登录是否成功
-                    if response.startswith("230 "):  # 230 表示登录成功
-                        print("Login successful.")
-                        logged_in = True
+                    email_command = input()
+
+                    if email_command.startswith("PASS "):
+                        response = send_command(sock, email_command)
+                        
+                        if response.startswith("230 "):  # 230 表示登录成功
+                            logged_in = True
+                        else:
+                            print("Login failed. Please try again.")
+                    elif email_command == 'QUIT' or email_command == 'ABOR':
+                        send_command(sock, 'QUIT')
+                        sock.close()
+                        break
                     else:
-                        print("Login failed. Please try again.")
+                        print("Invalid command. Please enter PASS command.")
                 else:
                     print("Invalid username. Please try again.")
+            elif user_command == 'QUIT' or user_command == 'ABOR':
+                send_command(sock, 'QUIT')
+                sock.close()
+                break
             else:
                 print("Invalid command. Please enter USER command first.")
 
         while logged_in:
-            command = input("Enter command: ")
+            command = input()
             if command.startswith("PORT"):
-                # 用户以 "127.0.0.1:8888" 格式提供IP和端口
-                ip_port = input("Enter your IP address and port (e.g., 127.0.0.1:8888): ")
+                if len(command.split(' ')) != 2:
+                    print("Amount of argument should be 1.")
+                    continue
+                ip_port = command.split(' ')[1]
                 try:
-                    ip_address, port_number = ip_port.split(':')
-                    port_number = int(port_number)
+                    ip_address, port_number_1, port_number_2 = ip_port.split(','[:4]), ip_port.split(',')[4], ip_port.split(',')[5]
+                    port_number_1 = int(port_number_1)
+                    port_number_2 = int(port_number_2)
 
                     # 检查端口号是否在有效范围内
-                    if 0 <= port_number <= 65535:
-                        ip_parts = ip_address.split('.')
-                        port1 = port_number // 256
-                        port2 = port_number % 256
-                        port_command = f"PORT {','.join(ip_parts)},{port1},{port2}"
-                        response = send_command(sock, port_command)
-
+                    if 0 <= 256 * port_number_1 + port_number_2 <= 65535:
                         data_connection_method.method = Method.PORT
-                        data_connection_method.ip_address = ip_address
-                        data_connection_method.port_number = port_number
+                        data_connection_method.ip_address = f"{ip_address[0]}.{ip_address[1]}.{ip_address[2]}.{ip_address[3]}"
+                        data_connection_method.port_number = 256 * port_number_1 + port_number_2
+
+                        port_command = f"PORT {ip_port}"
+                        data_connection_method.data_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+                        try:
+                            data_connection_method.data_sock.bind((data_connection_method.ip_address, data_connection_method.port_number))
+                        except socket.error as e:
+                            print(f"Failed to bind to {data_connection_method.ip_address}: \
+                                  {data_connection_method.port_number} - {e}", file=sys.stderr, flush=True)
+                            continue
+                        data_connection_method.data_sock.listen(1)
+
+                        response = send_command(sock, port_command)
                     else:
                         print("Invalid port number. Please enter a port number between 20000 and 65535.")
                 except ValueError:
@@ -335,7 +339,6 @@ def main():
                     parts = response.split('(')[1].split(')')[0].split(',')
                     data_ip_address = '.'.join(parts[:4])
                     data_port_number = int(parts[4]) * 256 + int(parts[5])
-                    print(f"Server IP: {data_ip_address}, Port: {data_port_number}")
 
                     # 创建数据连接
                     data_connection_method.method = Method.PASV
@@ -343,11 +346,17 @@ def main():
                     data_connection_method.port_number = data_port_number
 
             elif command.startswith("RETR"):
+                if len(command.split(' ')) != 2:
+                    print("Amount of argument should be 1.")
+                    continue
                 filename = command.split(' ')[1]
                 if retrieve_file(sock, filename, data_connection_method):
                     data_connection_method.method = Method.NOTHING
 
             elif command.startswith("STOR"):
+                if len(command.split(' ')) != 2:
+                    print("Amount of argument should be 1.")
+                    continue
                 filename = command.split(' ')[1]
                 if store_file(sock, filename, data_connection_method):
                     data_connection_method.method = Method.NOTHING
@@ -356,14 +365,16 @@ def main():
                 send_command(sock, "SYST")
 
             elif command.startswith("TYPE"):
-                type_code = input("Enter transfer type (A for ASCII, I for binary): ").strip().upper()
-                send_command(sock, f"TYPE {type_code}")
+                send_command(sock, command)
 
             elif command.startswith("LIST"):
                 if list_files(sock, data_connection_method):
                     data_connection_method.method = Method.NOTHING
 
             elif command.startswith("CWD"):
+                if len(command.split(' ')) != 2:
+                    print("Amount of argument should be 1.")
+                    continue
                 directory = command.split(' ')[1]
                 change_directory(sock, directory)
 
@@ -371,14 +382,20 @@ def main():
                 print_working_directory(sock)
 
             elif command.startswith("MKD"):
+                if len(command.split(' ')) != 2:
+                    print("Amount of argument should be 1.")
+                    continue
                 directory = command.split(' ')[1]
                 make_directory(sock, directory)
 
             elif command.startswith("RMD"):
+                if len(command.split(' ')) != 2:
+                    print("Amount of argument should be 1.")
+                    continue
                 directory = command.split(' ')[1]
                 remove_directory(sock, directory)
 
-            if command == 'QUIT' or command == 'ABOR':
+            elif command == 'QUIT' or command == 'ABOR':
                 send_command(sock, 'QUIT')
                 sock.close()
                 break
